% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{clear_cache}
\alias{clear_cache}
\title{Cache should be cleared at the begining of all imports to avoid sessions restarts to load changes.}
\usage{
clear_cache()
}
\description{
Cache should be cleared at the begining of all imports to avoid sessions restarts to load changes.
}

