% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{import - imports modules aloowing for relavtive path}
\usage{
import(relativePath)
}
\arguments{
\item{relativePath}{- string with relative path to the module}
}
\description{
import - imports modules aloowing for relavtive path
}

